+++ Sources/SourceGraph/SourceGraphMutatorRunner.swift
@@ -44,6 +44,7 @@ public final class SourceGraphMutatorRunner {
         PropertyWrapperRetainer.self,
         ResultBuilderRetainer.self,
         CodablePropertyRetainer.self,
+        EquatableHashablePropertyRetainer.self,
         ExternalOverrideRetainer.self,
 
         AncestralReferenceEliminator.self,